step to fix separate between numbers
