+++ lib/ui/components/extension_string_number_widget.dart
@@ -1,0 +1,6 @@
+extension StringNumberExtension on String {
+  String spaceSeparateNumbers() {
+    final result = replaceAllMapped(RegExp(r'(\d{1,3})(?=(\d{3})+(?!\d))'), (Match m) => '${m[1]} ');
+    return result;
+  }
+}
