+++ pubspec.yaml
@@ -49,6 +49,7 @@ dependencies:
   intl_phone_field: ^3.2.0
   dropdown_button2: ^2.3.9
   mask_text_input_formatter: ^2.5.0
+  intl: ^0.18.1
 
 dev_dependencies:
   flutter_test:
